maildir-notmuch-sync: pre/post subcommands, ext tagging script, cleanup
This patch splits maildir-notmuch-sync into two parts using a pre and
post subcommand, meant to be called from pre- and post-sync OfflineIMAP
hooks.

Additionally it fixes whitespace issues and typos.

It removes hard-coded assumptions about an "archive" directory and
replaces affected examples with "inbox" equivalents.

Finally it allows for an external initial tagging script to be called.
This part is a bit messy. In general the script should be refactored to
parse commandline args and options better.

Signed-off-by: Michael Turquette <mturquette@deferred.io>